Fixed a few things and going to fix few more.

+++ instrumenter/clean.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 
-rm -rf $1 $1.bkp 
 fileName=$(echo $1| cut -d '.' -f 1)
 rm -rf ${fileName}_instru.c
 
-mv $1.orig $1
+if [ -e $1.orig ]
+then 
+	rm -rf $1 $1.bkp 
+	mv $1.orig $1
+fi
